Fix async within persistent actors
To work with async workflows in persistent actors, the RunTask method
should be called instead of calling the ActorTaskScheduler directly.

Besides that, the task CE must be used instead of converting the async
to task using the Async.StartAsTask in order for it to work. For that
reason, the FSharp.Core was updated to > 6.0, since the task CE was
introduced in F# 6.